backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-multipart
beautifulsoup4
requests
Pillow
pytest
httpx
